+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # master
 
+* You can now have a collection of object at root node, you should use decodeCollection() if you don't use identified object
+
+# 1.1.0 - 2016-08-03
+
 * Add feature for typed json
 
 # 1.0.0 - 2016-04-26